Move resolveUsers to createClient
